batch of LLM added improvements

+++ src/models/indirect.cpp
@@ -41,8 +41,9 @@ const IndirectMemory* Indirect::GetMemory(
 void Indirect::Predict(ShortTermMemory& short_term_memory,
                        const LongTermMemory& long_term_memory) {
   const auto& m = *GetMemory(long_term_memory);
-  unsigned int context = ((context_ << 8) + short_term_memory.bit_context) %
-                         m.nonstationary_table.size();
+  unsigned int context =
+      ((static_cast<uint64_t>(context_) << 8) + short_term_memory.bit_context) %
+      m.nonstationary_table.size();
   int nonstationary_state = m.nonstationary_table[context];
   // 255 means this context has never been seen.
   if (nonstationary_state != 255) {
@@ -60,8 +61,9 @@ void Indirect::Predict(ShortTermMemory& short_term_memory,
 void Indirect::Learn(const ShortTermMemory& short_term_memory,
                      LongTermMemory& long_term_memory) {
   auto& m = *GetMemory(long_term_memory);
-  unsigned int context = ((context_ << 8) + short_term_memory.bit_context) %
-                         m.nonstationary_table.size();
+  unsigned int context =
+      ((static_cast<uint64_t>(context_) << 8) + short_term_memory.bit_context) %
+      m.nonstationary_table.size();
   int nonstationary_state = m.nonstationary_table[context];
   if (nonstationary_state == 255) {
     // 255 is the uninitialized state, so we the reset to a valid "0" state.
